change int to long int

+++ main.h
@@ -26,7 +26,7 @@ int _putchar(char c);
 int print_percent(va_list args);
 int print_char(va_list args);
 int print_string(va_list args);
-int print_int_helper(int n);
+int print_int_helper(long int n);
 int print_int(va_list args);
 int _printf(const char *format, ...);
